#85 [fix] getBearTypeUseCase를 이용해서 루틴 완료 뷰 띄우기
