+++ app/src/main/java/com/sopetit/softie/ui/dailyroutine/complete/DailyRoutineCompleteViewModel.kt
@@ -1,0 +1,16 @@
+package com.sopetit.softie.ui.dailyroutine.complete
+
+import androidx.lifecycle.ViewModel
+import com.sopetit.softie.domain.usecase.GetBearTypeUseCase
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class DailyRoutineCompleteViewModel @Inject constructor(
+    private val getBearTypeUseCase: GetBearTypeUseCase
+) : ViewModel() {
+
+    fun getBearType(): String {
+        return getBearTypeUseCase()
+    }
+}
